+++ spring-cloud-starter-tencent-polaris-circuitbreaker/src/main/java/com/tencent/cloud/polaris/circuitbreaker/reporter/SuccessCircuitBreakerReporter.java
@@ -19,17 +19,16 @@ package com.tencent.cloud.polaris.circuitbreaker.reporter;
 
 import java.util.Optional;
 
-import com.tencent.cloud.rpc.enhancement.AbstractPolarisReporterAdapter;
 import com.tencent.cloud.rpc.enhancement.config.RpcEnhancementReporterProperties;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPlugin;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginContext;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginType;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedRequestContext;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedResponseContext;
+import com.tencent.cloud.rpc.enhancement.plugin.PolarisEnhancedPluginUtils;
 import com.tencent.cloud.rpc.enhancement.plugin.reporter.SuccessPolarisReporter;
 import com.tencent.polaris.api.plugin.circuitbreaker.ResourceStat;
 import com.tencent.polaris.circuitbreak.api.CircuitBreakAPI;
-import com.tencent.polaris.client.api.SDKContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -38,17 +37,22 @@ import org.springframework.cloud.client.ServiceInstance;
 
 import static com.tencent.cloud.rpc.enhancement.plugin.PluginOrderConstant.ClientPluginOrder.CIRCUIT_BREAKER_REPORTER_PLUGIN_ORDER;
 
-
-public class SuccessCircuitBreakerReporter extends AbstractPolarisReporterAdapter implements EnhancedPlugin {
+/**
+ * SuccessCircuitBreakerReporter.
+ *
+ * @author sean yu
+ */
+public class SuccessCircuitBreakerReporter implements EnhancedPlugin {
 
 	private static final Logger LOG = LoggerFactory.getLogger(SuccessPolarisReporter.class);
 
 	private final CircuitBreakAPI circuitBreakAPI;
 
+	private final RpcEnhancementReporterProperties reportProperties;
+
 	public SuccessCircuitBreakerReporter(RpcEnhancementReporterProperties reportProperties,
-			SDKContext context,
 			CircuitBreakAPI circuitBreakAPI) {
-		super(reportProperties, context);
+		this.reportProperties = reportProperties;
 		this.circuitBreakAPI = circuitBreakAPI;
 	}
 
@@ -59,20 +63,20 @@ public class SuccessCircuitBreakerReporter extends AbstractPolarisReporterAdapte
 
 	@Override
 	public EnhancedPluginType getType() {
-		return EnhancedPluginType.POST;
+		return EnhancedPluginType.Client.POST;
 	}
 
 	@Override
 	public void run(EnhancedPluginContext context) throws Throwable {
-		if (!super.reportProperties.isEnabled()) {
+		if (!this.reportProperties.isEnabled()) {
 			return;
 		}
 		EnhancedRequestContext request = context.getRequest();
 		EnhancedResponseContext response = context.getResponse();
-		ServiceInstance serviceInstance = Optional.ofNullable(context.getServiceInstance())
+		ServiceInstance serviceInstance = Optional.ofNullable(context.getTargetServiceInstance())
 				.orElse(new DefaultServiceInstance());
 
-		ResourceStat resourceStat = createInstanceResourceStat(
+		ResourceStat resourceStat = PolarisEnhancedPluginUtils.createInstanceResourceStat(
 				serviceInstance.getServiceId(),
 				serviceInstance.getHost(),
 				serviceInstance.getPort(),