+++ spring-cloud-tencent-rpc-enhancement/src/main/java/com/tencent/cloud/rpc/enhancement/config/RpcEnhancementAutoConfiguration.java
@@ -24,20 +24,22 @@ import com.netflix.zuul.ZuulFilter;
 import com.tencent.cloud.polaris.context.ConditionalOnPolarisEnabled;
 import com.tencent.cloud.polaris.context.config.PolarisContextAutoConfiguration;
 import com.tencent.cloud.rpc.enhancement.feign.EnhancedFeignBeanPostProcessor;
+import com.tencent.cloud.rpc.enhancement.filter.EnhancedReactiveFilter;
+import com.tencent.cloud.rpc.enhancement.filter.EnhancedServletFilter;
 import com.tencent.cloud.rpc.enhancement.plugin.DefaultEnhancedPluginRunner;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPlugin;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginRunner;
 import com.tencent.cloud.rpc.enhancement.plugin.reporter.ExceptionPolarisReporter;
 import com.tencent.cloud.rpc.enhancement.plugin.reporter.SuccessPolarisReporter;
-import com.tencent.cloud.rpc.enhancement.resttemplate.BlockingLoadBalancerClientAspect;
 import com.tencent.cloud.rpc.enhancement.resttemplate.EnhancedRestTemplateInterceptor;
-import com.tencent.cloud.rpc.enhancement.resttemplate.RibbonLoadBalancerClientAspect;
+import com.tencent.cloud.rpc.enhancement.resttemplate.PolarisLoadBalancerRequestTransformer;
 import com.tencent.cloud.rpc.enhancement.scg.EnhancedGatewayGlobalFilter;
-import com.tencent.cloud.rpc.enhancement.webclient.EnhancedWebClientReporter;
+import com.tencent.cloud.rpc.enhancement.webclient.EnhancedWebClientExchangeFilterFunction;
 import com.tencent.cloud.rpc.enhancement.webclient.PolarisLoadBalancerClientRequestTransformer;
+import com.tencent.cloud.rpc.enhancement.webclient.RibbonLoadBalancerClientAspect;
 import com.tencent.cloud.rpc.enhancement.zuul.EnhancedErrorZuulFilter;
 import com.tencent.cloud.rpc.enhancement.zuul.EnhancedPostZuulFilter;
-import com.tencent.cloud.rpc.enhancement.zuul.EnhancedPreZuulFilter;
+import com.tencent.cloud.rpc.enhancement.zuul.EnhancedRouteZuulFilter;
 import com.tencent.polaris.api.core.ConsumerAPI;
 import com.tencent.polaris.client.api.SDKContext;
 
@@ -49,15 +51,25 @@ import org.springframework.boot.autoconfigure.AutoConfigureBefore;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.cloud.client.loadbalancer.LoadBalanced;
+import org.springframework.cloud.client.serviceregistry.Registration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.context.annotation.Role;
+import org.springframework.core.annotation.Order;
 import org.springframework.core.env.Environment;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.reactive.function.client.WebClient;
+
+import static javax.servlet.DispatcherType.ASYNC;
+import static javax.servlet.DispatcherType.ERROR;
+import static javax.servlet.DispatcherType.FORWARD;
+import static javax.servlet.DispatcherType.INCLUDE;
+import static javax.servlet.DispatcherType.REQUEST;
 
 /**
  * Auto Configuration for Polaris {@link feign.Feign} OR {@link RestTemplate} which can automatically bring in the call
@@ -73,23 +85,55 @@ import org.springframework.web.reactive.function.client.WebClient;
 public class RpcEnhancementAutoConfiguration {
 
 	@Bean
+	@Lazy
 	public EnhancedPluginRunner enhancedFeignPluginRunner(
-			@Autowired(required = false) List<EnhancedPlugin> enhancedPlugins) {
-		return new DefaultEnhancedPluginRunner(enhancedPlugins);
+			@Autowired(required = false) List<EnhancedPlugin> enhancedPlugins,
+			@Autowired(required = false) Registration registration,
+			SDKContext sdkContext) {
+		return new DefaultEnhancedPluginRunner(enhancedPlugins, registration, sdkContext);
 	}
 
 	@Bean
 	public SuccessPolarisReporter successPolarisReporter(RpcEnhancementReporterProperties properties,
-			SDKContext context,
 			ConsumerAPI consumerAPI) {
-		return new SuccessPolarisReporter(properties, context, consumerAPI);
+		return new SuccessPolarisReporter(properties, consumerAPI);
 	}
 
 	@Bean
 	public ExceptionPolarisReporter exceptionPolarisReporter(RpcEnhancementReporterProperties properties,
-			SDKContext context,
 			ConsumerAPI consumerAPI) {
-		return new ExceptionPolarisReporter(properties, context, consumerAPI);
+		return new ExceptionPolarisReporter(properties, consumerAPI);
+	}
+
+	@Configuration(proxyBeanMethods = false)
+	@ConditionalOnWebApplication(type = ConditionalOnWebApplication.Type.SERVLET)
+	protected static class RpcEnhancementServletFilterConfig {
+
+		@Bean
+		public FilterRegistrationBean<EnhancedServletFilter> enhancedServletFilterRegistrationBean(
+				EnhancedServletFilter enhancedServletFilter) {
+			FilterRegistrationBean<EnhancedServletFilter> filterRegistrationBean =
+					new FilterRegistrationBean<>(enhancedServletFilter);
+			filterRegistrationBean.setDispatcherTypes(ASYNC, ERROR, FORWARD, INCLUDE, REQUEST);
+			filterRegistrationBean.setOrder(enhancedServletFilter.getClass().getAnnotation(Order.class).value());
+			return filterRegistrationBean;
+		}
+
+		@Bean
+		public EnhancedServletFilter enhancedServletFilter(@Lazy EnhancedPluginRunner pluginRunner) {
+			return new EnhancedServletFilter(pluginRunner);
+		}
+
+	}
+
+	@Configuration(proxyBeanMethods = false)
+	@ConditionalOnWebApplication(type = ConditionalOnWebApplication.Type.REACTIVE)
+	protected static class RpcEnhancementReactiveFilterConfig {
+
+		@Bean
+		public EnhancedReactiveFilter enhancedReactiveFilter(@Lazy EnhancedPluginRunner pluginRunner) {
+			return new EnhancedReactiveFilter(pluginRunner);
+		}
 	}
 
 	/**
@@ -141,16 +185,9 @@ public class RpcEnhancementAutoConfiguration {
 
 		@Bean
 		@ConditionalOnMissingBean
-		@ConditionalOnClass(name = {"org.springframework.cloud.loadbalancer.blocking.client.BlockingLoadBalancerClient"})
-		public BlockingLoadBalancerClientAspect blockingLoadBalancerClientAspect() {
-			return new BlockingLoadBalancerClientAspect();
-		}
-
-		@Bean
-		@ConditionalOnMissingBean
-		@ConditionalOnClass(name = {"org.springframework.cloud.netflix.ribbon.RibbonLoadBalancerClient"})
-		public RibbonLoadBalancerClientAspect ribbonLoadBalancerClientAspect() {
-			return new RibbonLoadBalancerClientAspect();
+		@ConditionalOnClass(name = {"org.springframework.cloud.client.loadbalancer.LoadBalancerRequestTransformer"})
+		public PolarisLoadBalancerRequestTransformer polarisLoadBalancerRequestTransformer() {
+			return new PolarisLoadBalancerRequestTransformer();
 		}
 	}
 
@@ -166,12 +203,12 @@ public class RpcEnhancementAutoConfiguration {
 		private List<WebClient.Builder> webClientBuilder = Collections.emptyList();
 
 		@Bean
-		public EnhancedWebClientReporter exchangeFilterFunction(@Lazy EnhancedPluginRunner pluginRunner) {
-			return new EnhancedWebClientReporter(pluginRunner);
+		public EnhancedWebClientExchangeFilterFunction exchangeFilterFunction(@Lazy EnhancedPluginRunner pluginRunner) {
+			return new EnhancedWebClientExchangeFilterFunction(pluginRunner);
 		}
 
 		@Bean
-		public SmartInitializingSingleton addEnhancedWebClientReporterForWebClient(EnhancedWebClientReporter reporter) {
+		public SmartInitializingSingleton addEnhancedWebClientReporterForWebClient(EnhancedWebClientExchangeFilterFunction reporter) {
 			return () -> webClientBuilder.forEach(webClient -> {
 				webClient.filter(reporter);
 			});
@@ -184,6 +221,12 @@ public class RpcEnhancementAutoConfiguration {
 			return new PolarisLoadBalancerClientRequestTransformer();
 		}
 
+		@Bean
+		@ConditionalOnMissingBean
+		@ConditionalOnClass(name = {"org.springframework.cloud.netflix.ribbon.RibbonLoadBalancerClient"})
+		public RibbonLoadBalancerClientAspect ribbonLoadBalancerClientAspect() {
+			return new RibbonLoadBalancerClientAspect();
+		}
 	}
 
 	/**
@@ -211,8 +254,8 @@ public class RpcEnhancementAutoConfiguration {
 	@ConditionalOnClass(name = "com.netflix.zuul.http.ZuulServlet")
 	protected static class PolarisCircuitBreakerZuulFilterConfig {
 		@Bean
-		public EnhancedPreZuulFilter enhancedZuulPreFilter(@Lazy EnhancedPluginRunner pluginRunner, Environment environment) {
-			return new EnhancedPreZuulFilter(pluginRunner, environment);
+		public EnhancedRouteZuulFilter enhancedZuulRouteFilter(@Lazy EnhancedPluginRunner pluginRunner, Environment environment) {
+			return new EnhancedRouteZuulFilter(pluginRunner, environment);
 		}
 
 		@Bean