+++ spring-cloud-tencent-rpc-enhancement/src/main/java/com/tencent/cloud/rpc/enhancement/plugin/reporter/SuccessPolarisReporter.java
@@ -19,16 +19,15 @@ package com.tencent.cloud.rpc.enhancement.plugin.reporter;
 
 import java.util.Optional;
 
-import com.tencent.cloud.rpc.enhancement.AbstractPolarisReporterAdapter;
 import com.tencent.cloud.rpc.enhancement.config.RpcEnhancementReporterProperties;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPlugin;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginContext;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginType;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedRequestContext;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedResponseContext;
+import com.tencent.cloud.rpc.enhancement.plugin.PolarisEnhancedPluginUtils;
 import com.tencent.polaris.api.core.ConsumerAPI;
 import com.tencent.polaris.api.rpc.ServiceCallResult;
-import com.tencent.polaris.client.api.SDKContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -42,16 +41,17 @@ import static com.tencent.cloud.rpc.enhancement.plugin.PluginOrderConstant.Clien
  *
  * @author Haotian Zhang
  */
-public class SuccessPolarisReporter extends AbstractPolarisReporterAdapter implements EnhancedPlugin {
+public class SuccessPolarisReporter implements EnhancedPlugin {
 
 	private static final Logger LOG = LoggerFactory.getLogger(SuccessPolarisReporter.class);
 
 	private final ConsumerAPI consumerAPI;
 
-	public SuccessPolarisReporter(RpcEnhancementReporterProperties properties,
-			SDKContext context,
+	private final RpcEnhancementReporterProperties reportProperties;
+
+	public SuccessPolarisReporter(RpcEnhancementReporterProperties reportProperties,
 			ConsumerAPI consumerAPI) {
-		super(properties, context);
+		this.reportProperties = reportProperties;
 		this.consumerAPI = consumerAPI;
 	}
 
@@ -62,24 +62,25 @@ public class SuccessPolarisReporter extends AbstractPolarisReporterAdapter imple
 
 	@Override
 	public EnhancedPluginType getType() {
-		return EnhancedPluginType.POST;
+		return EnhancedPluginType.Client.POST;
 	}
 
 	@Override
 	public void run(EnhancedPluginContext context) {
-		if (!super.reportProperties.isEnabled()) {
+		if (!this.reportProperties.isEnabled()) {
 			return;
 		}
 
 		EnhancedRequestContext request = context.getRequest();
 		EnhancedResponseContext response = context.getResponse();
-		ServiceInstance serviceInstance = Optional.ofNullable(context.getServiceInstance())
-				.orElse(new DefaultServiceInstance());
+		ServiceInstance callerServiceInstance = Optional.ofNullable(context.getLocalServiceInstance()).orElse(new DefaultServiceInstance());
+		ServiceInstance calleeServiceInstance = Optional.ofNullable(context.getTargetServiceInstance()).orElse(new DefaultServiceInstance());
 
-		ServiceCallResult resultRequest = createServiceCallResult(
-				serviceInstance.getServiceId(),
-				serviceInstance.getHost(),
-				serviceInstance.getPort(),
+		ServiceCallResult resultRequest = PolarisEnhancedPluginUtils.createServiceCallResult(
+				callerServiceInstance.getHost(),
+				calleeServiceInstance.getServiceId(),
+				calleeServiceInstance.getHost(),
+				calleeServiceInstance.getPort(),
 				request.getUrl(),
 				request.getHttpHeaders(),
 				response.getHttpHeaders(),
@@ -89,8 +90,7 @@ public class SuccessPolarisReporter extends AbstractPolarisReporterAdapter imple
 		);
 
 		LOG.debug("Will report ServiceCallResult of {}. Request=[{} {}]. Response=[{}]. Delay=[{}]ms.",
-				resultRequest.getRetStatus().name(), request.getHttpMethod().name(), request.getUrl()
-						.getPath(), response.getHttpStatus(), context.getDelay());
+				resultRequest.getRetStatus().name(), request.getHttpMethod().name(), request.getUrl().getPath(), response.getHttpStatus(), context.getDelay());
 
 		consumerAPI.updateServiceCallResult(resultRequest);