+++ spring-cloud-tencent-rpc-enhancement/src/main/java/com/tencent/cloud/rpc/enhancement/scg/EnhancedGatewayGlobalFilter.java
@@ -21,6 +21,7 @@ import java.net.URI;
 
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginContext;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginRunner;
+import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginType;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedRequestContext;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedResponseContext;
 import reactor.core.publisher.Mono;
@@ -33,10 +34,6 @@ import org.springframework.cloud.gateway.route.Route;
 import org.springframework.core.Ordered;
 import org.springframework.web.server.ServerWebExchange;
 
-import static com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginType.EXCEPTION;
-import static com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginType.FINALLY;
-import static com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginType.POST;
-import static com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginType.PRE;
 import static org.springframework.cloud.gateway.support.ServerWebExchangeUtils.GATEWAY_REQUEST_URL_ATTR;
 import static org.springframework.cloud.gateway.support.ServerWebExchangeUtils.GATEWAY_ROUTE_ATTR;
 
@@ -64,8 +61,15 @@ public class EnhancedGatewayGlobalFilter implements GlobalFilter, Ordered {
 				.build();
 		enhancedPluginContext.setRequest(enhancedRequestContext);
 
+//		enhancedPluginContext.setLocalServiceInstance(pluginRunner.getLocalServiceInstance());
+//		Response<ServiceInstance> serviceInstanceResponse = exchange.getAttribute(GATEWAY_LOADBALANCER_RESPONSE_ATTR);
+//		if (serviceInstanceResponse != null && serviceInstanceResponse.hasServer()) {
+//			ServiceInstance instance = serviceInstanceResponse.getServer();
+//			enhancedPluginContext.setTargetServiceInstance(instance);
+//		}
+
 		// Run pre enhanced plugins.
-		pluginRunner.run(PRE, enhancedPluginContext);
+		pluginRunner.run(EnhancedPluginType.Client.PRE, enhancedPluginContext);
 
 		long startTime = System.currentTimeMillis();
 		return chain.filter(exchange)
@@ -80,7 +84,7 @@ public class EnhancedGatewayGlobalFilter implements GlobalFilter, Ordered {
 						serviceInstance.setHost(uri.getHost());
 						serviceInstance.setPort(uri.getPort());
 					}
-					enhancedPluginContext.setServiceInstance(serviceInstance);
+					enhancedPluginContext.setTargetServiceInstance(serviceInstance);
 				})
 				.doOnSuccess(v -> {
 					enhancedPluginContext.setDelay(System.currentTimeMillis() - startTime);
@@ -91,18 +95,18 @@ public class EnhancedGatewayGlobalFilter implements GlobalFilter, Ordered {
 					enhancedPluginContext.setResponse(enhancedResponseContext);
 
 					// Run post enhanced plugins.
-					pluginRunner.run(POST, enhancedPluginContext);
+					pluginRunner.run(EnhancedPluginType.Client.POST, enhancedPluginContext);
 				})
 				.doOnError(t -> {
 					enhancedPluginContext.setDelay(System.currentTimeMillis() - startTime);
 					enhancedPluginContext.setThrowable(t);
 
 					// Run exception enhanced plugins.
-					pluginRunner.run(EXCEPTION, enhancedPluginContext);
+					pluginRunner.run(EnhancedPluginType.Client.EXCEPTION, enhancedPluginContext);
 				})
 				.doFinally(v -> {
 					// Run finally enhanced plugins.
-					pluginRunner.run(FINALLY, enhancedPluginContext);
+					pluginRunner.run(EnhancedPluginType.Client.FINALLY, enhancedPluginContext);
 				});
 	}