+++ spring-cloud-tencent-rpc-enhancement/src/main/java/com/tencent/cloud/rpc/enhancement/zuul/EnhancedErrorZuulFilter.java
@@ -24,22 +24,18 @@ import com.netflix.zuul.ZuulFilter;
 import com.netflix.zuul.context.RequestContext;
 import com.netflix.zuul.exception.ZuulException;
 import com.tencent.cloud.common.constant.ContextConstant;
-import com.tencent.cloud.common.util.ZuulFilterUtils;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginContext;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginRunner;
+import com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginType;
 import com.tencent.cloud.rpc.enhancement.plugin.EnhancedResponseContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import org.springframework.cloud.client.DefaultServiceInstance;
-import org.springframework.cloud.netflix.ribbon.apache.RibbonApacheHttpResponse;
 import org.springframework.core.env.Environment;
 import org.springframework.http.HttpHeaders;
 import org.springframework.util.StringUtils;
 
 import static com.tencent.cloud.common.constant.ContextConstant.Zuul.POLARIS_PRE_ROUTE_TIME;
-import static com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginType.EXCEPTION;
-import static com.tencent.cloud.rpc.enhancement.plugin.EnhancedPluginType.FINALLY;
 import static org.springframework.cloud.netflix.zuul.filters.support.FilterConstants.ERROR_TYPE;
 
 /**
@@ -90,13 +86,9 @@ public class EnhancedErrorZuulFilter extends ZuulFilter {
 			enhancedPluginContext = (EnhancedPluginContext) enhancedPluginContextObj;
 		}
 
-		DefaultServiceInstance serviceInstance = new DefaultServiceInstance();
-		Object ribbonResponseObj = context.get("ribbonResponse");
 		Object startTimeMilliObject = context.get(POLARIS_PRE_ROUTE_TIME);
 		Throwable throwable = context.getThrowable();
-		RibbonApacheHttpResponse ribbonResponse;
-		if (throwable != null && ribbonResponseObj != null && ribbonResponseObj instanceof RibbonApacheHttpResponse
-				&& startTimeMilliObject != null && startTimeMilliObject instanceof Long) {
+		if (throwable != null && startTimeMilliObject != null && startTimeMilliObject instanceof Long) {
 			HttpHeaders responseHeaders = new HttpHeaders();
 			Collection<String> names = context.getResponse().getHeaderNames();
 			for (String name : names) {
@@ -110,17 +102,12 @@ public class EnhancedErrorZuulFilter extends ZuulFilter {
 			Long startTimeMilli = (Long) startTimeMilliObject;
 			enhancedPluginContext.setDelay(System.currentTimeMillis() - startTimeMilli);
 			enhancedPluginContext.setThrowable(throwable);
-			ribbonResponse = (RibbonApacheHttpResponse) ribbonResponseObj;
-			serviceInstance.setServiceId(ZuulFilterUtils.getServiceId(context));
-			serviceInstance.setHost(ribbonResponse.getRequestedURI().getHost());
-			serviceInstance.setPort(ribbonResponse.getRequestedURI().getPort());
-			enhancedPluginContext.setServiceInstance(serviceInstance);
 
 			// Run post enhanced plugins.
-			pluginRunner.run(EXCEPTION, enhancedPluginContext);
+			pluginRunner.run(EnhancedPluginType.Client.EXCEPTION, enhancedPluginContext);
 
 			// Run finally enhanced plugins.
-			pluginRunner.run(FINALLY, enhancedPluginContext);
+			pluginRunner.run(EnhancedPluginType.Client.FINALLY, enhancedPluginContext);
 		}
 		return null;
 	}